requirements.txt
numpy
keras
tensorflow
scipy
imageio
streamlit
Pillow


